requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
pyspark>=3.4.0
sentence-transformers>=2.2.2
pinecone-client>=2.2.1
python-dotenv>=1.0.0
scikit-learn>=1.2.0
pulp>=2.7.0 